# openapi.yaml
openapi: 3.0.3
info:
  title: MVTMS REST API
  description: |
    OpenAPI specification for the MVTMS REST API v1. Authentication is via
    a signed Bearer token (`Authorization: Bearer <token>`) or an existing
    session cookie. The token endpoints are public; all other `/api/v1/*`
    routes require a valid principal plus Casbin RBAC permissions.
  version: "1.0.0"
  contact:
    name: MVTMS Support
    url: https://avandab.com
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/vehicles:
    get:
      tags: [Vehicles]
      summary: List vehicles
      description: Requires `vehicles:read` permission. Supports SOP fleet_class (CV/PV/FS/OFS) and ownership (O/C/F/M) filters.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
        - name: fleet_class
          in: query
          schema:
            type: string
            enum: [CV, PV, FS, OFS]
        - name: ownership
          in: query
          schema:
            type: string
            enum: [O, C, F, M]
      responses:
        '200':
          description: Paginated list of vehicles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Vehicles]
      summary: Register a vehicle (SOP IE31)
      description: Requires `vehicles:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '201':
          description: Vehicle registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed API token returned by `POST /api/v1/auth/token`.
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based).
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Items per page.
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search filter.
    StatusParam:
      name: status
      in: query
      schema:
        type: string
      description: Status filter.
  schemas:
...
    IDResponse:
      type: object
      properties:
        id:
          type: string
...
    VehicleProfile:
      type: object
      description: SOP fleet-object master data (IE31 tabs + Vehicle Master).
      properties:
        fleet_class:
          type: string
          enum: [CV, PV, FS, OFS]
        ownership:
          type: string
          enum: [O, C, F, M]
        fleet_number:
          type: string
        description:
          type: string
        manufacturer:
          type: string
        manuf_country:
          type: string
        model:
          type: string
        constr_year_month:
          type: string
        acquisition_value:
          type: number
        acquisition_currency:
          type: string
        acquisition_date:
          type: string
          format: date
        purchase_vendor:
          type: string
        valid_from:
          type: string
          format: date
        valid_to:
          type: string
          format: date
        facility_id:
          type: string
        maint_plant:
          type: string
        planning_plant:
          type: string
        company_code:
          type: string
        business_area:
          type: string
        cost_center:
          type: string
        asset_no:
          type: string
        fleet_object_no:
          type: string
        chassis_no:
          type: string
        vehicle_category:
          type: string
        engine_number:
          type: string
        engine_power:
          type: string
        engine_capacity:
          type: string
        cylinder_count:
          type: integer
        max_speed:
          type: number
        weight:
          type: number
        weight_unit:
          type: string
          enum: [TO, KG]
        load_volume:
          type: number
        volume_unit:
          type: string
        secondary_fuel:
          type: string
        usage_indicator:
          type: string
    Vehicle:
      type: object
      properties:
        id:
          type: string
        registration_number:
          type: string
        vehicle_number:
          type: string
        vehicle_type:
          type: string
        capacity:
          type: integer
        fuel_type:
          type: string
        insurance_expiry:
          type: string
          format: date
        fitness_expiry:
          type: string
          format: date
        permit_expiry:
          type: string
          format: date
        status:
          type: string
        current_mileage:
          type: number
          nullable: true
        profile:
          $ref: '#/components/schemas/VehicleProfile'
    VehicleList:
      type: object
      properties:
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        total:
          type: integer
          format: int64
    CreateVehicleRequest:
      type: object
      required: [registration_number, vehicle_number, insurance_expiry, fitness_expiry, permit_expiry]
      properties:
        registration_number:
          type: string
        vehicle_number:
          type: string
        vehicle_type:
          type: string
        capacity:
          type: integer
        fuel_type:
          type: string
        insurance_expiry:
          type: string
          format: date
        fitness_expiry:
          type: string
          format: date
        permit_expiry:
          type: string
          format: date
        current_mileage:
          type: number
        profile:
          $ref: '#/components/schemas/VehicleProfile'
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient RBAC permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'